Extracting out sp1 config system into its own folder, adding new comments and re-aranging code, down to 11 FIXMEs
