+++ nori/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod bridge_head;
 pub mod sp1_prover;
+pub mod sp1_prover_config;
 pub mod utils;
 pub mod rpcs;
 pub mod contract;
